refactor: create new service for distribute students
